# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SQU Weather Station API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SensorMeta {
  id: string;
  displayName: string;
  csvFile: string;
  description: string;
  parameterCount: number;
  color: string;
}

export interface SensorSummary {
  id: string;
  displayName: string;
  description: string;
  color: string;
  latestTimestamp: string;
  parameterCount: number;
  mainParameter: string;
  mainValue: string;
  mainUnit: string;
  rowCount: number;
}

export type SensorReadingData = { [key: string]: unknown };

export interface SensorReading {
  timestamp: string;
  data: SensorReadingData;
}

export type SensorDataResponseRowsItem = { [key: string]: unknown };

export interface SensorDataResponse {
  sensorId: string;
  total: number;
  rows: SensorDataResponseRowsItem[];
}

export interface ParameterMeta {
  key: string;
  label: string;
  unit: string;
  category: string;
  iconType: string;
}

export interface LoggerSettings {
  intervalLabel: string;
  intervalCode: string;
  updatedAt?: string;
  message?: string;
}

export interface SetIntervalBody {
  intervalLabel: string;
  intervalCode: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetSensorDataParams = {
  /**
   * Start datetime (YYYY-MM-DDTHH:mm)
   */
  from?: string;
  /**
   * End datetime (YYYY-MM-DDTHH:mm)
   */
  to?: string;
  /**
   * Max rows to return (default 500)
   */
  limit?: number;
};

export type ExportExcelParams = {
  /**
   * Sensor ID or 'all'
   */
  sensor?: string;
  from?: string;
  to?: string;
};